+++ backend/server/src/api.rs
@@ -1,0 +1,13 @@
+use types::{errors::ApiError, lang::Language};
+
+pub mod editing;
+pub mod public;
+pub mod testing;
+
+pub(crate) fn parse_language(lang: &str) -> Result<Language, ApiError> {
+    match lang {
+        "sv" => Ok(Language::Sv),
+        "en" => Ok(Language::En),
+        _ => Err(ApiError::BadRequest("Invalid language".to_string())),
+    }
+}