+++ backend/server/src/api/editing.rs
@@ -1,0 +1,5 @@
+pub mod chapters;
+pub mod courses;
+pub mod prefixes;
+pub mod problems;
+pub mod topics;