+++ backend/server/src/api/editing/prefixes.rs
@@ -1,7 +1,7 @@
 use axum::{Json, extract::Path, http::StatusCode, response::IntoResponse};
 use serde_json::json;
 
-use crate::{PrefixEntry, prefixes};
+use db::{PrefixEntry, prefixes};
 use types::errors::ApiError;
 
 /// Returns all the data about every prefix in the DB as a `Vec<PrefixEntry>`.