+++ backend/server/src/api/public.rs
@@ -1,0 +1,5 @@
+mod course_contents;
+mod translations;
+
+pub use course_contents::*;
+pub use translations::*;